fix: same as b4

+++ app/backend/src/entities/TravelPlans/RawTravelPlan.entity.cr
@@ -2,5 +2,5 @@ class RawTravelPlan
   include JSON::Serializable
 
   @[JSON::Field]
-  property id : Int64
+  property id : Int32
 end
